fix(gx): fix incorrect cast in CGxString::InitializeViewTranslation

+++ src/gx/font/CGxString.cpp
@@ -580,7 +580,7 @@ void CGxString::InitializeViewTranslation() {
         this->m_viewTranslation.x += (this->m_blockWidth * 0.5f);
     }
 
-    float v13 = CMath::fuint_pi(this->m_spacing * static_cast<float>(GetScreenPixelHeight())) / GetScreenPixelHeight();
+    float v13 = CMath::fuint_pi(this->m_spacing * static_cast<float>(GetScreenPixelHeight())) / static_cast<float>(GetScreenPixelHeight());
     float v24 = v13 * (float)(this->m_intB0 - 1) + (float)this->m_intB0 * this->m_currentFontHeight;
 
     if (this->m_intD4 >= 4) {
